Move the `PartialEq` and `Eq` impls for arrays to a separate file
